requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.0
python-dotenv==1.0.0
pdfplumber==0.10.3
nltk==3.8.1
textblob==0.17.1
